search

+++ app/src/main/java/com/sweetmay/advancedcryptoindicators2/model/entity/coin/detailed/MarketCap.kt
@@ -1,7 +1,7 @@
 package com.sweetmay.advancedcryptoindicators2.model.entity.coin.detailed
 
 data class MarketCap(
-    val btc: Int,
+    val btc: Double,
     val rub: Long,
     val usd: Long,
 )
